+++ kinto/core/permission/memory.py
@@ -1,10 +1,11 @@
 import re
 
 from kinto.core.permission import PermissionBase
+from kinto.core.utils import synchronized
 
 
 class Permission(PermissionBase):
-    """Permission backend implementation in local thread memory.
+    """Permission backend implementation in local process memory.
 
     Enable in configuration::
 
@@ -24,12 +25,14 @@ class Permission(PermissionBase):
     def flush(self):
         self._store = {}
 
+    @synchronized
     def add_user_principal(self, user_id, principal):
         user_key = 'user:%s' % user_id
         user_principals = self._store.get(user_key, set())
         user_principals.add(principal)
         self._store[user_key] = user_principals
 
+    @synchronized
     def remove_user_principal(self, user_id, principal):
         user_key = 'user:%s' % user_id
         user_principals = self._store.get(user_key, set())
@@ -43,6 +46,7 @@ class Permission(PermissionBase):
         else:
             self._store[user_key] = user_principals
 
+    @synchronized
     def remove_principal(self, principal):
         for user_principals in self._store.values():
             try:
@@ -50,17 +54,20 @@ class Permission(PermissionBase):
             except KeyError:
                 pass
 
+    @synchronized
     def get_user_principals(self, user_id):
         user_key = 'user:%s' % user_id
         members = self._store.get(user_key, set())
         return members
 
+    @synchronized
     def add_principal_to_ace(self, object_id, permission, principal):
         permission_key = 'permission:%s:%s' % (object_id, permission)
         object_permission_principals = self._store.get(permission_key, set())
         object_permission_principals.add(principal)
         self._store[permission_key] = object_permission_principals
 
+    @synchronized
     def remove_principal_from_ace(self, object_id, permission, principal):
         permission_key = 'permission:%s:%s' % (object_id, permission)
         object_permission_principals = self._store.get(permission_key, set())
@@ -74,11 +81,13 @@ class Permission(PermissionBase):
         else:
             self._store[permission_key] = object_permission_principals
 
+    @synchronized
     def get_object_permission_principals(self, object_id, permission):
         permission_key = 'permission:%s:%s' % (object_id, permission)
         members = self._store.get(permission_key, set())
         return members
 
+    @synchronized
     def get_accessible_objects(self, principals, bound_permissions=None):
         principals = set(principals)
         candidates = []
@@ -101,12 +110,14 @@ class Permission(PermissionBase):
                 perms_by_object_id.setdefault(object_id, set()).add(perm)
         return perms_by_object_id
 
+    @synchronized
     def get_authorized_principals(self, bound_permissions):
         principals = set()
         for obj_id, perm in bound_permissions:
             principals |= self.get_object_permission_principals(obj_id, perm)
         return principals
 
+    @synchronized
     def get_objects_permissions(self, objects_ids, permissions=None):
         result = []
         for object_id in objects_ids:
@@ -124,6 +135,7 @@ class Permission(PermissionBase):
             result.append(perms)
         return result
 
+    @synchronized
     def replace_object_permissions(self, object_id, permissions):
         for permission, principals in permissions.items():
             permission_key = 'permission:%s:%s' % (object_id, permission)
@@ -133,6 +145,7 @@ class Permission(PermissionBase):
                 self._store[permission_key] = set(principals)
         return permissions
 
+    @synchronized
     def delete_object_permissions(self, *object_id_list):
         to_delete = []
         for key in self._store.keys():